completely replace x509 extension parsing.
The old extension code did not work (the x509_extension event
was never raised). It also did not include any data that could be further
processed.

The new event returns the data as it would be displayed by the openssl
commmand line client.

Examples:
[name=X509v3 Extended Key Usage, short_name=extendedKeyUsage, oid=2.5.29.37, critical=F, value=TLS Web Server Authentication, TLS Web Client Authentication]

[name=X509v3 Basic Constraints, short_name=basicConstraints, oid=2.5.29.19, critical=T, value=CA:TRUE]

This commit should probably end up in 2.2, no matter if the rest of the
X509 parsing changes gets in.